refactor: Consolidate store availability tests into algoliaLocalizedProduct test suite
